+++ backoffice/src/routes/web.php
@@ -436,6 +436,17 @@ return function (App $app, Twig $twig): void {
         return $controller->bulkSetTipologieIoService($request, $response);
     });
 
+    // Backup / Import configurazione
+    $app->post('/configurazione/backup/export', function($request, $response) use ($twig) {
+        $controller = new ConfigurazioneController($twig);
+        return $controller->exportBackup($request, $response);
+    });
+
+    $app->post('/configurazione/backup/import', function($request, $response) use ($twig) {
+        $controller = new ConfigurazioneController($twig);
+        return $controller->importBackup($request, $response);
+    });
+
     $app->get('/users/new', function($request, $response) use ($twig) {
         if (isset($_SESSION['user'])) {
             $twig->getEnvironment()->addGlobal('current_user', $_SESSION['user']);